Some micro-optimization work in the invoker
A few changes here, mostly the purpose of this change is to avoid async
state machines and Task<T> when not needed. This should extrafastpath the
fast path when you have no filters for a given stage. YES, I am talking
just about executing the compiler-generated async preamble, about the case
where we go async - AsyncTaskMethodBuilder.Start is about 5% (inclusive)
of execution time inside MVC/Routing.

Also some micro-optimization inside GetNextFilter. The impact of this is
hard to quantify because it's sometimes inlined. Changes here try to hit
the early-out case with only a single cast, and avoid doing the array
access twice. This saved a few instructions in the IL.